Implement RedisDB#Publish() and Miniredis#Publish()
refs #1
